Move extensions out of the main library
